+++ src/invoices/InvoicesList/InvoicesListContainer.js
@@ -44,6 +44,7 @@ const buildInvoicesQuery = makeQueryBuilder(
     [FILTERS.INVOICE_DATE]: buildDateRangeQuery.bind(null, [FILTERS.INVOICE_DATE]),
     [FILTERS.PAYMENT_DUE]: buildDateRangeQuery.bind(null, [FILTERS.PAYMENT_DUE]),
     [FILTERS.APPROVAL_DATE]: buildDateTimeRangeQuery.bind(null, [FILTERS.APPROVAL_DATE]),
+    [FILTERS.PAYMENT_DATE]: buildDateTimeRangeQuery.bind(null, [FILTERS.PAYMENT_DATE]),
     [FILTERS.TAGS]: (filterValue) => {
       const value = Array.isArray(filterValue) ? filterValue.join('" or "') : filterValue;