+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 
+* Display and filter paymentDate in Invoice module. Refs UINV-224.
 * Make invoice line 'vendor reference number' and type repeatable, paired fields and populate from POL. Refs UINV-165.
 * Show the possible duplicate modal on invoice edit. Refs UINV-209.
 * Use `CurrencyExchangeRateFields` component from stripes-acq-components. Refs UIOR-610.